+++ packages/core/src/components/Tipseen/Tipseen.tsx
@@ -56,7 +56,7 @@ export interface TipseenProps extends VibeComponentProps {
    */
   modifiers?: Array<Modifier<unknown>>;
   closeAriaLabel?: string;
-  onClose?: () => void;
+  onClose?: (event?: React.MouseEvent<HTMLButtonElement>) => void;
   content: ElementContent;
   /**
    * Control the color of the Tipseen close button. Dark theme can be useful while presenting bright images under the tipseen image
@@ -184,6 +184,7 @@ const Tipseen: VibeComponent<TipseenProps> & {
           showDelay={showDelay}
           hideTrigger={hideTrigger}
           showTrigger={showTrigger}
+          showOnDialogEnter={false}
           content={tooltipContent}
           theme={color === "inverted" ? "dark" : "primary"}
           containerSelector={containerSelector}